fix(show-me/desktopcapturer): refactor to use current Electron security defaults (#898)
