+++ static/show-me/desktopcapturer/index.html
@@ -5,8 +5,5 @@
 </head>
 <body>
   <video></video>
-  <script>
-    require('./renderer.js')
-  </script>
 </body>
 </html>
