+++ static/show-me/desktopcapturer/main.js
@@ -5,13 +5,16 @@
 // https://electronjs.org/docs/api/desktop-capturer
 
 const { app, BrowserWindow } = require('electron')
+const path = require('path')
 
 app.whenReady().then(() => {
   const mainWindow = new BrowserWindow({
     height: 600,
     width: 600,
     webPreferences: {
-      nodeIntegration: true
+      nodeIntegration: false, // default in Electron >= 5
+      contextIsolation: true, // default in Electron >= 12
+      preload: path.join(__dirname, 'preload.js')
     }
   })
